Prime cache is user data dir is empty (#54)
* add basic version that inserts a single script snippet in the root html file

* spacing fix

* inject critical requests extraction code in server

* add console message extraction

* add critical request extraction code that should be added inside server.py

* add flag for extract critical requests

* add command line option

for passing in a chrome user-data-dir

* change default value of user-data-dir

* parse arguments to alter behavior

- user-data-dir is no longer ignored
- extract critical requests is no longer ignored

* initial work on speed index

add base js files to determine speed index of a page

* send the output file argument to pwmetrics

* update pwmetrics dependency version

* use pwmetrics to get speed index

* add python wrapper for speed index

now we an control whether speed index is computed or not through a CLI option

* make user-data-dir optional

instead of a compulsory string

* inject script in all html

not just root

* output the har to stdout

this was just a change i made for debugging. made it into a commit by mistake.

* pwmetrics does not need some parameters

i was previously checking if user-data-dir was included, but this is optional

* speedindex saving to outputfile 

fixed the logic: previously i was having a temp file and then moving it to the desired output. now i write directly to the output file requested.

* fix speed index code

works with optional user-data-dir and prints json to screen correctly

* change function param name

from har to speed index

* remove js file as this is embedded within the python code

* implement support for extracting critical request in devtools

and start working on double running to get har diff

* support for extracting critical requests in replay

* mahimahi now passes critical requests param

* injecting critcal extractor is now based on parameters

previously was always

* capture_har now supports optional critical extractor

* capturer returns har with critical urls only when the flag is set

* recorder should return har/si based on input params

* add critical flag in har definition

* finish implementing speed index

still need to figure out why the SI value is much larger than PLT

* add critical attribute to HAR

and change some pylint config

* add critical attribute to har, add some docstrings

* fix tests for capture har

* add basic test for missing argument exception

* add support for speed index in push.py

* pass in chrome user-dir as docker mount

* user-data-dir chrome parameter format changed

* linting issue

extra newline and line length

* add support for critical requests in preprocess

* add support for critical requests in python

* fix test

* prime the cache if empty user-data-dir is passed in